# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LandOverSEA - International Dating & Cultural Connection API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /discover/cards:
    get:
      operationId: getDiscoverCards
      tags: [discover]
      summary: Get a batch of profiles to swipe on
      parameters:
        - name: limit
          in: query
          schema:
            type: number
            default: 20
        - name: gender
          in: query
          schema:
            type: ["string", "null"]
        - name: minAge
          in: query
          schema:
            type: ["number", "null"]
        - name: maxAge
          in: query
          schema:
            type: ["number", "null"]
        - name: countries
          in: query
          schema:
            type: ["string", "null"]
            description: Comma-separated country codes
        - name: globalMode
          in: query
          schema:
            type: boolean
            default: true
        - name: languages
          in: query
          schema:
            type: ["string", "null"]
        - name: offset
          in: query
          schema:
            type: number
            default: 0
        - name: verifiedOnly
          in: query
          schema:
            type: boolean
        - name: relationshipGoal
          in: query
          schema:
            type: ["string", "null"]
        - name: longDistance
          in: query
          schema:
            type: boolean
        - name: relocation
          in: query
          schema:
            type: boolean
        - name: interestsOverlap
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DiscoverCard"
  /discover/swipe:
    post:
      operationId: swipe
      tags: [discover]
      summary: Like, pass, or superlike a profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SwipeInput"
      responses:
        "200":
          description: OK — includes match info if mutual
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SwipeResult"
        "400":
          description: Invalid target, action, or self-swipe
        "403":
          description: A block exists between the users
        "404":
          description: Target user not found
  /discover/filters:
    get:
      operationId: getDiscoverFilters
      tags: [discover]
      summary: Get current discover filter settings
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiscoverFilters"
    patch:
      operationId: updateDiscoverFilters
      tags: [discover]
      summary: Update discover filter settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiscoverFiltersInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiscoverFilters"

  # ── MATCHES ───────────────────────────────────────────────────────────────
...
  /premium/who-liked-me:
    get:
      tags: [premium]
      operationId: getWhoLikedMe
      summary: Get users who liked the current user (premium only)
      responses:
        '200':
          description: List of profiles who liked you
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PublicProfile'
        '403':
          description: Premium required
components:
  schemas:
...
    ProfilePhoto:
      type: object
      required: [id, url, position, isPrimary]
      properties:
        id:
          type: number
          multipleOf: 1
        url:
          type: string
        position:
          type: number
          multipleOf: 1
        isPrimary:
          type: boolean
...
    PublicProfile:
      type: object
      required: [id, userId, name]
      properties:
        id:
          type: number
        userId:
          type: number
        name:
          type: string
        age:
          type: ["number", "null"]
        bio:
          type: ["string", "null"]
        occupation:
          type: ["string", "null"]
        height:
          type: ["number", "null"]
        gender:
          type: ["string", "null"]
        primaryLanguage:
          type: ["string", "null"]
        otherLanguages:
          type: array
          items:
            type: string
        country:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        culturalInterests:
          type: array
          items:
            type: string
        interests:
          type: array
          items:
            type: string
        relationshipGoal:
          type: ["string", "null"]
        photos:
          type: array
          items:
            $ref: "#/components/schemas/ProfilePhoto"
        voiceIntroUrl:
          type: ["string", "null"]
        isVerified:
          type: boolean
        role:
          type: ["string", "null"]
          description: Participant role within a group conversation (owner, admin, member)
...
    DiscoverCard:
      type: object
      required: [userId, profile]
      properties:
        userId:
          type: number
        profile:
          $ref: "#/components/schemas/PublicProfile"
        distanceKm:
          type: ["number", "null"]
        sharedInterests:
          type: array
          items:
            type: string
        sharedLanguages:
          type: array
          items:
            type: string
    SwipeInput:
      type: object
      required: [targetUserId, action]
      properties:
        targetUserId:
          type: number
        action:
          type: string
          enum: [like, pass, superlike]
        idempotencyKey:
          type: ["string", "null"]
    SwipeResult:
      type: object
      required: [action]
      properties:
        action:
          type: string
        isMatch:
          type: boolean
        match:
          $ref: "#/components/schemas/Match"
        conversationId:
          type: ["number", "null"]
    DiscoverFilters:
      type: object
      required:
        [
          minAge,
          maxAge,
          gender,
          globalMode,
          interestsOverlap,
          verifiedOnly,
          longDistance,
          relocation,
        ]
      properties:
        minAge:
          type: number
        maxAge:
          type: number
        gender:
          type: ["string", "null"]
          enum: [male, female, non_binary, null]
        countries:
          type: array
          items:
            type: string
        languages:
          type: array
          items:
            type: string
        globalMode:
          type: boolean
        relationshipGoal:
          type: ["string", "null"]
        interestsOverlap:
          type: boolean
        verifiedOnly:
          type: boolean
        longDistance:
          type: boolean
        relocation:
          type: boolean
    DiscoverFiltersInput:
      type: object
      properties:
        minAge:
          type: number
        maxAge:
          type: number
        gender:
          type: ["string", "null"]
        countries:
          type: array
          items:
            type: string
        languages:
          type: array
          items:
            type: string
        globalMode:
          type: boolean
        relationshipGoal:
          type: ["string", "null"]
        interestsOverlap:
          type: boolean
        verifiedOnly:
          type: boolean
        longDistance:
          type: boolean
        relocation:
          type: boolean

    # ── MATCHES ───────────────────────────────────────────────────────────
    Match:
      type: object
      required: [id, userId1, userId2, createdAt]
      properties:
        id:
          type: number
        userId1:
          type: number
        userId2:
          type: number
        conversationId:
          type: ["number", "null"]
        otherUser:
          $ref: "#/components/schemas/PublicProfile"
        createdAt:
          type: string
          format: date-time
...